refactor(LegFinder): Move to otp package.
